+++ docs/pyagentspec/source/changelog.rst
@@ -4,6 +4,25 @@ Changelog
 Agent Spec |release|
 --------------------
 
+New features
+^^^^^^^^^^^^
+
+Improvements
+^^^^^^^^^^^^
+
+* ** Added back the CrewAI adapter**
+
+  The CrewAI adapter has been added back.
+
+Bug fixes
+^^^^^^^^^
+
+Breaking Changes
+^^^^^^^^^^^^^^^^
+
+Agent Spec 26.3.0
+-----------------
+
 Improvements
 ^^^^^^^^^^^^
 
@@ -60,6 +79,10 @@ New features
 
 Breaking Changes
 ^^^^^^^^^^^^^^^^
+
+* **Removed CrewAI adapter**
+
+  The CrewAI adapter and its optional dependencies have been removed from ``pyagentspec`` due to a CVE in a pinned dependency of CrewAI (Chroma DB 1.1.0) at the time of our release.
 
 
 Agent Spec 26.1.2